# swagger.yaml
openapi: 3.0.3
info:
  title: PetPulse AI Backend Services
  description: Complete API collection for PetPulse AI backend services including Auth and User modules
  version: 1.0.0
  contact:
    name: PetPulse AI Team
    email: team@petpulseai.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.petpulseai.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    Role:
      type: string
      enum:
        - SUPER_ADMIN
        - ADMIN
        - NORMAL
      description: User role in the system
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@example.com"
        name:
          type: string
          description: User full name
          example: "John Doe"
        image:
          type: string
          format: uri
          description: User profile image URL
          example: "https://example.com/avatar.jpg"
          nullable: true
        signUpMethod:
          type: string
          description: Sign up method used
          example: "email"
          default: "google"
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          description: User creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: User last update timestamp
      required:
        - id
        - email
        - name
        - role
        - createdAt
        - updatedAt
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "newuser@example.com"
        name:
          type: string
          description: User full name
          example: "New User"
        password:
          type: string
          minLength: 6
          description: User password
          example: "securePassword123"
        image:
          type: string
          format: uri
          description: User profile image URL
          example: "https://example.com/avatar.jpg"
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
          default: "NORMAL"
      required:
        - email
        - name
        - password
    SigninRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "test@example.com"
        password:
          type: string
          description: User password
          example: "securePassword123"
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
      required:
        - access_token
        - user
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Error message
          example: "Validation failed"
        error:
          type: string
          description: Error type
          example: "Bad Request"
      required:
        - statusCode
        - message
...
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Sign Up
      description: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            examples:
              normal_user:
                summary: Normal User Registration
                value:
                  email: "newuser@example.com"
                  name: "New User"
                  password: "securePassword123"
                  image: "https://example.com/avatar.jpg"
                  role: "NORMAL"
              admin_user:
                summary: Admin User Registration
                value:
                  email: "admin@example.com"
                  name: "Admin User"
                  password: "adminPassword123"
                  role: "ADMIN"
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signin:
    post:
      tags:
        - Authentication
      summary: Sign In
      description: Authenticate user and get access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
      responses:
        '200':
          description: User successfully signed in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...